# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /imports:
    get:
      operationId: listImports
      tags: [imports]
      summary: List legacy workbook imports
      responses:
        "200":
          description: Import history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ImportSummary"
    post:
      operationId: createImport
      tags: [imports]
      summary: Save a validated import review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportInput"
      responses:
        "201":
          description: Import review saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportSummary"
  /imports/{importId}:
    get:
      operationId: getImport
      tags: [imports]
      summary: Get import review details
      parameters:
        - $ref: "#/components/parameters/ImportId"
      responses:
        "200":
          description: Import review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportDetail"
        "404":
          description: Import not found
  /imports/{importId}/commit:
    post:
      operationId: commitImport
      tags: [imports]
      summary: Commit an immutable legacy baseline
      parameters:
        - $ref: "#/components/parameters/ImportId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommitImportInput"
      responses:
        "200":
          description: Baseline committed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportDetail"
        "409":
          description: Import cannot be committed
        "404":
          description: Import not found
...
components:
  parameters:
    ImportId:
      name: importId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    QualityIssue:
      type: object
      required: [code, severity, title, detail, rowNumbers]
      properties:
        code:
          type: string
        severity:
          type: string
          enum: [error, warning, review]
        title:
          type: string
        detail:
          type: string
        rowNumbers:
          type: array
          items:
            type: integer
    PreviewRow:
      type: object
      required: [sourceRow, systemId, identifier, nomenclature, unit, currentDglp, currentEchs]
      properties:
        sourceRow:
          type: integer
        systemId:
          type: ["string", "null"]
        identifier:
          type: ["string", "null"]
        nomenclature:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        currentDglp:
          type: ["number", "null"]
        currentEchs:
          type: ["number", "null"]
    ImportInput:
      type: object
      required: [sourceFileName, sourceFileHash, worksheetName, rowCount, columnCount, departmentCount, uniqueIdentifierCount, missingIdentifierCount, conflictGroupCount, warnings, errors, previewRows]
      properties:
        sourceFileName:
          type: string
        sourceFileHash:
          type: string
        fileSize:
          type: integer
        worksheetName:
          type: string
        rowCount:
          type: integer
        columnCount:
          type: integer
        departmentCount:
          type: integer
        uniqueIdentifierCount:
          type: integer
        missingIdentifierCount:
          type: integer
        conflictGroupCount:
          type: integer
        warnings:
          type: integer
        errors:
          type: integer
        previewRows:
          type: array
          items:
            $ref: "#/components/schemas/PreviewRow"
        issues:
          type: array
          items:
            $ref: "#/components/schemas/QualityIssue"
        legacyRows:
          type: array
          items:
            $ref: "#/components/schemas/LegacyRowInput"
        departments:
          type: array
          items:
            $ref: "#/components/schemas/DepartmentInput"
        sourceObjectPath:
          type: ["string", "null"]
        sourceObjectContentType:
          type: ["string", "null"]
    CommitImportInput:
      type: object
      properties:
        acknowledged:
          type: boolean
    ImportSummary:
      type: object
      required: [id, sourceFileName, worksheetName, rowCount, departmentCount, status, createdAt, quality]
      properties:
        id:
          type: string
        sourceFileName:
          type: string
        worksheetName:
          type: string
        rowCount:
          type: integer
        columnCount:
          type: integer
        departmentCount:
          type: integer
        status:
          type: string
          enum: [review, committed]
        createdAt:
          type: string
        committedAt:
          type: ["string", "null"]
        quality:
          type: object
          required: [warnings, errors, conflicts, missingIdentifiers]
          properties:
            warnings:
              type: integer
            errors:
              type: integer
            conflicts:
              type: integer
            missingIdentifiers:
              type: integer
    ImportDetail:
      allOf:
        - $ref: "#/components/schemas/ImportSummary"
        - type: object
          required: [sourceFileHash, fileSize, uniqueIdentifierCount, previewRows, issues]
          properties:
            sourceFileHash:
              type: string
            fileSize:
              type: integer
            uniqueIdentifierCount:
              type: integer
            previewRows:
              type: array
              items:
                $ref: "#/components/schemas/PreviewRow"
            issues:
              type: array
              items:
                $ref: "#/components/schemas/QualityIssue"
...
    DepartmentInput:
      type: object
      required: [name, sourceColumnStart]
      properties:
        name:
          type: string
        sourceColumnStart:
          type: integer
    LegacyRowInput:
      type: object
      required: [sourceRow, sourceValues, departments]
      properties:
        sourceRow:
          type: integer
        systemId:
          type: ["string", "null"]
        identifier:
          type: ["string", "null"]
        nomenclature:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        previousPvmsMmf:
          type: ["number", "null"]
        currentPvmsMmf:
          type: ["number", "null"]
        previousDglpMmf:
          type: ["number", "null"]
        currentDglpMmf:
          type: ["number", "null"]
        previousEchsMmf:
          type: ["number", "null"]
        currentEchsMmf:
          type: ["number", "null"]
        lpr:
          type: ["string", "null"]
        sourceValues:
          type: array
          items: {}
        departments:
          type: array
          items:
            type: object
            required: [departmentIndex]
            properties:
              departmentIndex:
                type: integer
              pvms:
                type: ["string", "null"]
              dglp:
                type: ["number", "null"]
              echs:
                type: ["number", "null"]
...