# lib/api-zod/src/generated/types/importInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DepartmentInput } from './departmentInput';
import type { LegacyRowInput } from './legacyRowInput';
import type { PreviewRow } from './previewRow';
import type { QualityIssue } from './qualityIssue';

export interface ImportInput {
  sourceFileName: string;
  sourceFileHash: string;
  fileSize?: number;
  worksheetName: string;
  rowCount: number;
  columnCount: number;
  departmentCount: number;
  uniqueIdentifierCount: number;
  missingIdentifierCount: number;
  conflictGroupCount: number;
  warnings: number;
  errors: number;
  previewRows: PreviewRow[];
  issues?: QualityIssue[];
  legacyRows?: LegacyRowInput[];
  departments?: DepartmentInput[];
  /** @nullable */
  sourceObjectPath?: string | null;
  /** @nullable */
  sourceObjectContentType?: string | null;
}