# lib/api-zod/src/generated/types/legacyRowInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LegacyRowInputDepartmentsItem } from './legacyRowInputDepartmentsItem';

export interface LegacyRowInput {
  sourceRow: number;
  /** @nullable */
  systemId?: string | null;
  /** @nullable */
  identifier?: string | null;
  /** @nullable */
  nomenclature?: string | null;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  previousPvmsMmf?: number | null;
  /** @nullable */
  currentPvmsMmf?: number | null;
  /** @nullable */
  previousDglpMmf?: number | null;
  /** @nullable */
  currentDglpMmf?: number | null;
  /** @nullable */
  previousEchsMmf?: number | null;
  /** @nullable */
  currentEchsMmf?: number | null;
  /** @nullable */
  lpr?: string | null;
  sourceValues: unknown[];
  departments: LegacyRowInputDepartmentsItem[];
}